update splitting into some views, epic 1 update (not finished), table still shown in epic 1
