+++ GymBro/ViewModels/Epic2ViewModel.swift
@@ -1,11 +1,5 @@
-//
-//  Epic2ViewModel.swift
-//  cobacoba
-//
-//  Created by Benedick Wijayaputra on 16/07/24.
-//
-
-import SwiftUI
+import Foundation
+import Combine
 
 class Epic2ViewModel: ObservableObject {
     @Published var weight: String = ""
@@ -13,10 +7,33 @@ class Epic2ViewModel: ObservableObject {
     @Published var selectedSegment = 0
     @Published var result: Double?
     @Published var trainingWeights: [TrainingWeight] = []
+
+    @Published var oneRepMax: String = ""
+    private var cancellables = Set<AnyCancellable>()
+    private let calculator = OneRepMaxCalculator()
     
     let segments = ["1 Rep Max", "Training Weight"]
     let percentages: [Double] = [100, 95, 90, 85, 80, 75, 70]
     let repetitions: [Double] = [1, 2, 4, 6, 8, 10, 12]
+    
+    init() {
+        $weight
+            .combineLatest($reps)
+            .map { weight, reps in
+                self.calculateOneRepMax(weight: weight, reps: reps)
+            }
+            .assign(to: \.oneRepMax, on: self)
+            .store(in: &cancellables)
+    }
+    
+    private func calculateOneRepMax(weight: String, reps: String) -> String {
+        guard let weight = Double(weight), let reps = Int(reps), reps > 0 else {
+            return "Invalid input"
+        }
+        let result = calculator.calculateOneRepMax(weight: weight, reps: reps)
+        self.result = result
+        return String(format: "%.2f", result)
+    }
     
     func calculate() {
         guard let weightValue = Double(weight) else {
